Combined signers.rs
