+++ src/lib.rs
@@ -13,8 +13,7 @@ pub mod encode;
 mod fee_spec;
 pub mod keychain;
 pub mod psbt_ext;
-pub mod sd_card_signer;
-pub mod xkey_signer;
+pub mod signers;
 pub use fee_spec::*;
 
 pub use chacha20::cipher;
